requirements.txt
asgiref==3.5.0
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==4.0.2
django-cors-headers==3.11.0
django-extra-fields==3.0.2
django-heroku==0.3.1
django-js-asset==2.0.0
django-mptt==0.13.4
django-rest-auth==0.9.5
django-rest-passwordreset==1.2.1
django-rest-swagger==2.2.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
dnspython==2.2.0
drf-yasg==1.20.0
et-xmlfile==1.1.0
eventlet==0.33.0
falcon==3.0.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.1.0
mongoengine==0.24.0
numpy==1.22.3
openapi-codec==1.3.2
openpyxl==3.0.9
packaging==21.3
pandas==1.4.2
Pillow==9.0.1
psycopg2==2.9.3
psycopg2-binary==2.8.6
PyJWT==2.3.0
pymongo==4.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
whitenoise==6.0.0
